requirements.txt
beautifulsoup4==4.10.0
dash==2.6.2
multipledispatch==0.6.0
nltk==3.6.5
numpy==1.20.3
pandas==1.4.4
plotly==5.10.0
python-dateutil==2.8.2
pyyaml==6.0
requests==2.26.0
scikit-learn==1.1.1
selenium==4.5.0
tqdm==4.62.3
